refactor: Update logger.ts to send error messages to LINE Notify
This commit updates the logger.ts file to send error messages to LINE Notify in addition to logging them to the console. It creates an instance of the LINENotifyService and uses it to post the error message to the configured LINE Notify token. If there is an error sending the message, it is logged to the console.

Refactor the logger.ts file to include error message notifications via LINE Notify.